refactor folders, fix router
