+++ client/assets/favicon.ico
