+++ client/index.js
@@ -1,8 +1,9 @@
 import { h, render } from 'preact';
 import { Provider } from 'unistore/preact';
-import Routes from './routes';
+import { Router } from 'preact-router';
 
-import createStore from './store/store';
+import createStore from '../common/src/store/store';
+import App from '../common/src/App';
 
 const store = createStore(window.__STATE__);
 
@@ -10,7 +11,7 @@ const app = document.getElementById('app');
 
 render(
   <Provider store={store}>
-    <Routes />
+    <App/>
   </Provider>,
   app,
   app.lastChild
