+++ common/src/components/note/index.js
