+++ common/src/components/note/style.css
