+++ common/src/components/snip/style.css
