+++ common/src/manifest.json
